+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
+    "javascript.updateImportsOnFileMove.enabled": "prompt",
     "markdownlint.config": {
         "MD028": false,
         "MD025": {
