+++ packages/css-doc-page/src/tostaditas-components/hero/Hero.module.scss
