+++ packages/css-doc-page/src/views/docs/buttonDocs/ButtonDocs.jsx
@@ -1,16 +1,16 @@
 import React , { useState } from 'react';
-import ctaMark from '../../assets/markdown/buttons/cta.md';
-import ctaMarkIcon from '../../assets/markdown/buttons/cta-icon.md';
-import fillMark from '../../assets/markdown/buttons/primary.md';
-import fillMarkIcon from '../../assets/markdown/buttons/primary-icon.md';
-import textMark from '../../assets/markdown/buttons/text.md';
-import textMarkIcon from '../../assets/markdown/buttons/text-icon.md';
-import outlineMark from '../../assets/markdown/buttons/outline.md';
-import outlineMarkIcon from '../../assets/markdown/buttons/outline-icon.md';
-import iconMark from '../../assets/markdown/buttons/icon.md';
-import ButtonsIcon from '../../assets/img/button-docs.svg';
-import Markdown from '../../components/markdown';
-import Hero from '../../components/hero';
+import ctaMark from '../../../assets/markdown/buttons/cta.md';
+import ctaMarkIcon from '../../../assets/markdown/buttons/cta-icon.md';
+import fillMark from '../../../assets/markdown/buttons/primary.md';
+import fillMarkIcon from '../../../assets/markdown/buttons/primary-icon.md';
+import textMark from '../../../assets/markdown/buttons/text.md';
+import textMarkIcon from '../../../assets/markdown/buttons/text-icon.md';
+import outlineMark from '../../../assets/markdown/buttons/outline.md';
+import outlineMarkIcon from '../../../assets/markdown/buttons/outline-icon.md';
+import iconMark from '../../../assets/markdown/buttons/icon.md';
+import ButtonsIcon from '../../../assets/img/button-docs.svg';
+import Markdown from '../../../tostaditas-components/markdown';
+import Hero from '../../../tostaditas-components/hero';
 import styles from './ButtonDocs.module.scss';
 
 const Sources = {
