+++ packages/css-doc-page/src/views/guidelines/colorGuidelines/index.js
