+++ packages/css-doc-page/src/views/guidelines/typhographyGuidelines/TypographyGuidelines.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './TypographyGuidelines.module.scss';
-import Hero from '../../components/hero';
-import typography from '../../assets/img/type-thin.svg';
+import Hero from '../../../tostaditas-components/hero';
+import typography from '../../../assets/img/type-thin.svg';
 
 function TypographyGuidelines() {
   return (
