+++ packages/css-doc-page/src/views/index.js
@@ -1,13 +1,15 @@
-export { default as AccessibilitySpecifications } from './AccessibilitySpecifications'
-export { default as ButtonDocs } from './ButtonDocs'
-export { default as CheckboxDocs } from './CheckboxDocs'
+export { default as Home } from './Home'
+
+export { default as GettingStarted } from './GettingStarted'
 export { default as ColorGuidelines } from './ColorGuidelines'
 export { default as ComponentOverview } from './ComponentOverview'
 export { default as DesignGuidelines } from './DesignGuidelines'
-export { default as GettingStarted } from './GettingStarted'
-export { default as Home } from './Home'
 export { default as ShapeGuidelines } from './ShapeGuidelines'
 export { default as TokenPrinciples } from './TokenPrinciples'
 export { default as TyphographyGuidelines } from './TyphographyGuidelines'
-export { default as InputDocs } from './InputDocs'
-export { default as SelectDocs } from './SelectDocs'
+export { default as AccessibilitySpecifications } from './AccessibilitySpecifications'
+// Component Docs
+export { default as InputDocs } from './docs/inputDocs'
+export { default as SelectDocs } from './docs/selectDocs'
+export { default as ButtonDocs } from './docs/buttonDocs'
+export { default as CheckboxDocs } from './docs/checkboxDocs'
